Improve directory path comparisons
Properly deal with path separators when determining project base
directory when invoking Git hooks.

Properly ignore case on win32.  On darwin, the file system can be
either case sensitive or insensitive, with the latter being the
default.  Fortunately, it appears the casing is internally consistent
on that platform as we have not seen an issue on it.

Add both directories to skipping message in handlePushBasedEventOnRepo
to make it more clear what is going on.  In other words, I am not
certain the above changes will fix the issue on win32.

Move expandedTreeUtils test under parallel directory as source.

Update TypeScript config and package scripts.  Add test coverage.
Remove tsfmt.

Fixes #195