Fix directory inclusion testing on win32
The "belt & suspenders" commit.  When Git hooks were invoked, the
`$PWD` was a Cygwin-style path, e.g., C:\Some\Path was /c/Some/path.
This broke determination of whether a repo directory was under the
parent directory, which used process.cwd() which returned the C:\
style paths.  This was fixed in two ways.  First, on win32 we
lowercase the paths and compare and if that does not work we
cygwin-ize them and compare again.  Second, we no longer have the Git
hook shell script provide the current directory or any other
argument, preferring to do all that work in Node.js.  We do accept
values for branch and SHA, but any provided directory is ignored.

Use HookEvent enum type more strictly.

Fixes #195