Have post-receive hook script read from stdin
The post-receive Git hook script was not forwarding its standard input
to the `atomist git-hook post-receive` function it was calling.
Rather than fixing this omission, revert back to having the script
read the references from standard input and pass those as command-line
arguments to `atomist git-hook post-receive`.  This is better because
it allows other functionality in the post-receive script, should
people be using it.  The reading by the shell script was improved to
account for the case when multiple heads are pushed, calling `atomist
git-hook post-receive` for each one.

For future reference, the post-receive hook is called when a reference
from a clone of the repo is pushed to the parent repo, something that
happens, for example, when running transforms and autofixes in local
mode.  For each head (branch) pushed, the post-receive hook script
receives one line on standard input with the before SHA, after SHA,
and reference name, in that order and separated by white space.
https://git-scm.com/book/en/v2/Customizing-Git-Git-Hooks

Closes #210